+++ R/sensitivityplot.R
@@ -6,7 +6,8 @@
 sensitivityplot <- function(extendedres, BMDtype = c("zSD", "xfold"),
                             group, ECDF_plot = TRUE, colorby,
                             BMDsummary = c("first.quartile", "median" , "median.and.IQR"),
-                            BMD_log_transfo = TRUE)
+                            BMD_log_transfo = TRUE,
+                            line.size = 0.5, line.alpha = 0.5, point.alpha = 0.5)
 {
   BMDtype <- match.arg(BMDtype, c("zSD", "xfold"))
   BMDsummary <- match.arg(BMDsummary, c("first.quartile", "median", "median.and.IQR"))
@@ -94,11 +95,11 @@ sensitivityplot <- function(extendedres, BMDtype = c("zSD", "xfold"),
     } else
     {  
       gg <- ggplot(dnb, aes(x = .data$groupby, y = .data$firstquartile, 
-                             color = .data$level, alpha = I(0.8),
+                             color = .data$level, 
                              size = .data$nb_of_items))
     }
     
-    gg <- gg + geom_point(stat = 'identity')  +  coord_flip() +
+    gg <- gg + geom_point(stat = 'identity', alpha = point.alpha)  +  coord_flip() +
       labs(x = "", y = "BMD 25th quantiles") 
     
   } else {
@@ -113,18 +114,20 @@ sensitivityplot <- function(extendedres, BMDtype = c("zSD", "xfold"),
                                size = .data$nb_of_items))
       } else {
         gg <- ggplot(dnb, aes(x = .data$groupby, y = .data$secondquartile, 
-                               color = .data$level, alpha = I(0.5),
+                               color = .data$level, 
                                size = .data$nb_of_items))
       }
     }
-    gg <- gg + geom_point(stat = 'identity') + coord_flip() 
+    gg <- gg + geom_point(stat = 'identity', alpha = point.alpha) + coord_flip() 
     
     if (BMDsummary == "median")
     {
       gg <- gg + labs(x = "", y = "BMD medians") 
     } else {
       gg <- gg + geom_errorbar(aes(ymin = .data$firstquartile, 
-                                    ymax = .data$thirdquartile, linewidth = I(1)), 
+                                    ymax = .data$thirdquartile),
+                               linewidth = line.size, 
+                               alpha = line.alpha,
                                width = 0) +
         # line to remove lines on the size legend
         guides(size = guide_legend(override.aes = list(linetype = 0))) +